Hermitian and Symmetric Matrices

Hermitian

RandomMatrix.randHermitian — Function

randHermitian(d, n; diag, norm )

randHermitian(n; norm)

d : entry distribution

n : dimensions

norm : default is set to false, if norm set to true, then the matrix will be normalized with $n^{-1/2}$.

diag : the distribution for diagonal entries, by default diag=d. To use a different distribution (say Binomial) for digonal elements, set diag = Binomial(1,0.5)

Examples

Generates a 2 by 2 random Hermitian matrix with entries from the Standard Complex Gaussian.

randHermitian(2)

Generate a random 2 by 2 Symmetric Matrix with entries Poisson(2) rvs. Need to import the Distributions package for Poisson(2)

randHermitian(Poisson(2),2)
# or equivalently 
randSymmetric(Poisson(2),2)

# Entries uniformly from {1,2,3,...,10}
randHermitian(1:10,2)

# Entries either -1 or pi with equal probability
randHermitian([-1,pi],2)

Symmetric

RandomMatrix.randSymmetric — Function

randSymmetric(d::D, n::Int; Diag = d::D,  norm = false::Bool) where T<:S

randSymmetric(n::Int; norm = false::Bool)

d : entry distribution

n : dimensions

norm : default is set to false, if norm set to true, then the matrix will be normalized with n^(-1/2).

Diag : the distribution for diagonal entries, by default diag=d. To use a different distribution (say Binomial) for digonal elements, set diag = Binomial(1,0.5)

# Examples

# Generates a 2 by 2 random Symmetric matrix with entries from the Standard Gaussian.
randSymmetric(2)

Some Random Matrix Theory