RandomMatrix.randEllipic — Function

randEllipic(d::D, n::Int; r = 0.5::Float64, Diag=d::T, norm = false::Bool) where  D<:S

randEllipic(n::Int; r=0.5::Float64, norm = false::Bool)

d : default Normal(), entry distribution

n : dimensions

r : default 0.5, the correlation of $H_{ij},H_{ji}$ pairs

norm : default false, if norm set to true, then the matrix will be normalized with n^(-1/2).

Diag : default Diag=d, the distribution for diagonal entries. To use a different distribution (say Binomial) for digonal elements, set Diag = Binomial(1,0.5)

# Examples

# Generate a normalized random elliptic matrix, with correlation 0.5
A = randEllipic(500,norm=true)

# Plot the eigenvalues of A, this should look like an ellipse
A|>eigvals|>scatter

# Generate a normalized random elliptic matrix, with entries `Poisson(10)` correlation 0.1
using Distributions
randEllipic(Poisson(10),500, r=0.1 , norm=true)

RandomMatrix.resolvent — Function

resolvent(A::Matrix)

return the resolvent ***function*** of A, $R(z)= (A-zI)^{-1}$

# Examples

# This returns the resolvent **function** of a 500 by 500 Hermitian 
resolvent(randHermitian(500, norm = true)) # this is a generic function

# One can use the above return as a function
f = resolvent(randHermitian(500, norm = true))
f(0+0.1im) # this returns a matrix

RandomMatrix.ComplexNormal — Type

ComplexNormal(μ,σ)

μ : mean, μ = 0 by default

σ : standard deviation, σ =1 by default

# Examples

# Generates 10 iid standard Complex Gaussian r.v.s
rand(ComplexNormal(), 10) 

# Generates complex normal with mean 1+1im, variance 4
rand(ComplexNormal(1+1im,2)) 

RandomMatrix.Gaussian — Type

Gaussian(beta,μ,σ)

beta : 1 (default) for Real Gaussian, 2 for Complex Gaussian

μ : mean, μ = 0 by default

σ : standard deviation, σ =1 by default

# Generates 10 iid standard normal r.v.s
rand(Gaussian(), 10) 

# Generates a complex normal with mean 1+1im, variance 4
rand(Gaussian(2,1+1im,2))

RandomMatrix.:± — Function

±(a,b)

returns (a-b,a+b)

# Examples

1 ± 0.5  # returns (0.5,1.5)