RandomMatrix.randHermitian — Function

randHermitian(d::D, n::Int; Diag = d::D, norm = false::Bool) where T<:S

randHermitian(n::Int; norm = false::Bool)

d : entry distribution

n : dimensions

norm : default is set to false, if norm set to true, then the matrix will be normalized with n^(-1/2).

Diag : the distribution for diagonal entries, by default Diag=d. To use a different distribution (say Binomial) for digonal elements, set Diag = Binomial(1,0.5)

# Examples

# Generates a 2 by 2 random Hermitian matrix with entries from the Standard Complex Gaussian.
randHermitian(2)

# Generate a random 2 by 2 Symmetric Matrix with entries  `Poisson(2)` rvs. 
# *Need to import the `Distributions` package for `Poisson(2)`*
randHermitian(Poisson(2),2)
# or equivalently 
randSymmetric(Poisson(2),2)

# Entries uniformly from {1,2,3,...,10}
randHermitian(1:10,2)

# Entries either -1 or pi with equal probability
randHermitian([-1,pi],2)

RandomMatrix.randSymmetric — Function

randSymmetric(d::D, n::Int; Diag = d::D,  norm = false::Bool) where T<:S

randSymmetric(n::Int; norm = false::Bool)

d : entry distribution

n : dimensions

norm : default is set to false, if norm set to true, then the matrix will be normalized with n^(-1/2).

Diag : the distribution for diagonal entries, by default Diag=d. To use a different distribution (say Binomial) for digonal elements, set Diag = Binomial(1,0.5)

# Examples

# Generates a 2 by 2 random Symmetric matrix with entries from the Standard Gaussian.
randSymmetric(2)

RandomMatrix.randUnitary — Function

randUnitary(n::Int)

Generates a n by n random Unitary matrix

Equivalent to run rand(Haar(2,n))

For orthogonal matrices, use randOrthogonal or rand(Haar(1,n)) instead

# Examples

# Generate a 3 by 3 random Unitary matrix 
randUnitary(3)
# or
rand(Haar(2,3)) 

RandomMatrix.randOrthogonal — Function

randOrthogonal(n::Int)

Generates a n by n random Orthogonal matrix

Equivalent to run rand(Haar(1,n))

For unitary matrices, use randUnitary or rand(Haar(2,n)) instead

# Examples

# Generates a 3 by 3 random Orthogonal matrix 
randOrthogonal(3)
# or
rand(Haar(1,3))

RandomMatrix.Haar — Type

Haar(beta,n)

Uniform distribution on O(n) (beta = 1), U(n) (beta = 2)

beta: 1 for Orthogonal, 2 for Unitary

n: dimension

# Examples

# Generate a 100 by 100 random Unitary Matrix uniformly from U(n)
rand(Haar(2,100))

# Generate a 100 by 100 random Orthogonal Matrix uniformly from O(n)
rand(Haar(1,100))

RandomMatrix.randPermutation — Function

randPermutation(n; fix) 

n : dimension

fix : a keyword argument, default is set to fix = 0. If fix = x, randPermutation(n,x) will have atleast x fixed points

# Examples 

# Generates a random 5 by 5 permutation matrix
randPermutation(5)

# Generates  Generates a random 100 by 100 permutation matrix with atleast 10 fix points
randPermutation(100,fix=10)

RandomMatrix.randEllipic — Function

randEllipic(d::D, n::Int; r = 0.5::Float64, Diag=d::T, norm = false::Bool) where  D<:S

randEllipic(n::Int; r=0.5::Float64, norm = false::Bool)

d : default Normal(), entry distribution

n : dimensions

r : default 0.5, the correlation of $H_{ij},H_{ji}$ pairs

norm : default false, if norm set to true, then the matrix will be normalized with n^(-1/2).

Diag : default Diag=d, the distribution for diagonal entries. To use a different distribution (say Binomial) for digonal elements, set Diag = Binomial(1,0.5)

# Examples

# Generate a normalized random elliptic matrix, with correlation 0.5
A = randEllipic(500,norm=true)

# Plot the eigenvalues of A, this should look like an ellipse
A|>eigvals|>scatter

# Generate a normalized random elliptic matrix, with entries `Poisson(10)` correlation 0.1
using Distributions
randEllipic(Poisson(10),500, r=0.1 , norm=true)

RandomMatrix.resolvent — Function

resolvent(A::Matrix)

return the resolvent ***function*** of A, $R(z)= (A-zI)^{-1}$

# Examples

# This returns the resolvent **function** of a 500 by 500 Hermitian 
resolvent(randHermitian(500, norm = true)) # this is a generic function

# One can use the above return as a function
f = resolvent(randHermitian(500, norm = true))
f(0+0.1im) # this returns a matrix

RandomMatrix.ComplexNormal — Type

ComplexNormal(μ,σ)

μ : mean, μ = 0 by default

σ : standard deviation, σ =1 by default

# Examples

# Generates 10 iid standard Complex Gaussian r.v.s
rand(ComplexNormal(), 10) 

# Generates complex normal with mean 1+1im, variance 4
rand(ComplexNormal(1+1im,2)) 

RandomMatrix.Gaussian — Type

Gaussian(beta,μ,σ)

beta : 1 (default) for Real Gaussian, 2 for Complex Gaussian

μ : mean, μ = 0 by default

σ : standard deviation, σ =1 by default

# Generates 10 iid standard normal r.v.s
rand(Gaussian(), 10) 

# Generates a complex normal with mean 1+1im, variance 4
rand(Gaussian(2,1+1im,2))

RandomMatrix.MarchenkoPastur — Type

MarchenkoPastur(λ,σ)

Marchenko-Pastur r.v. with asymptotic ratio λ and scale parameter σ.

λ by default is 0.5

σ by default is 1, the value 1 for σ is typically used in Random Matrix Theory.

pdf(d::MarchenkoPastur,x::Real)  

# Examples 

# Generate a MP rv with ρ = 0.5, σ =1
rand(MarchenkoPastur()) 

# Generate a 100 by 100 matrix with entries i.i.d MP rvs with ρ=0.2
rand(MarchenkoPastur(0.2),100,100)

# Generate 100 MP rvs with ρ=0.1 and σ=2
rand(MarchenkoPastur(0.1,2),100)

# Compute the desity for the MP distribution with λ=1.6 at the point 0
pdf(MarchenkoPastur(1.6),0)

RandomMatrix.:± — Function

±(a,b)

returns (a-b,a+b)

# Examples

1 ± 0.5  # returns (0.5,1.5)